# Apply select clippy recommendations