--- a/csv-core/src/reader.rs
+++ b/csv-core/src/reader.rs
@@ -439,7 +439,7 @@ enum NfaState {
 }
 
 /// A list of NFA states that have an explicit representation in the DFA.
-const NFA_STATES: &'static [NfaState] = &[
+const NFA_STATES: &[NfaState] = &[
     NfaState::StartRecord,
     NfaState::StartField,
     NfaState::EndFieldDelim,
@@ -1238,7 +1238,7 @@ impl DfaClasses {
             panic!("added too many classes")
         }
         self.classes[b as usize] = self.next_class as u8;
-        self.next_class = self.next_class + 1;
+        self.next_class += 1;
     }
 
     fn num_classes(&self) -> usize {