--- a/src/error.rs
+++ b/src/error.rs
@@ -228,7 +228,7 @@ pub struct FromUtf8Error {
 impl FromUtf8Error {
     /// Create a new FromUtf8Error.
     pub(crate) fn new(rec: ByteRecord, err: Utf8Error) -> FromUtf8Error {
-        FromUtf8Error { record: rec, err: err }
+        FromUtf8Error { record: rec, err }
     }
 
     /// Access the underlying `ByteRecord` that failed UTF-8 validation.
@@ -271,7 +271,7 @@ pub struct Utf8Error {
 
 /// Create a new UTF-8 error.
 pub fn new_utf8_error(field: usize, valid_up_to: usize) -> Utf8Error {
-    Utf8Error { field: field, valid_up_to: valid_up_to }
+    Utf8Error { field, valid_up_to }
 }
 
 impl Utf8Error {
@@ -315,7 +315,7 @@ impl<W> IntoInnerError<W> {
     /// (This is a visibility hack. It's public in this module, but not in the
     /// crate.)
     pub(crate) fn new(wtr: W, err: io::Error) -> IntoInnerError<W> {
-        IntoInnerError { wtr: wtr, err: err }
+        IntoInnerError { wtr, err }
     }
 
     /// Returns the error which caused the call to `into_inner` to fail.