using clippy v0.1.57 for the following lints:
* [redundant_static_lifetimes](https://rust-lang.github.io/rust-clippy/master/index.html#redundant_static_lifetimes)
* [redundant_field_names](https://rust-lang.github.io/rust-clippy/master/index.html#redundant_field_names)
* [assign_op_pattern](https://rust-lang.github.io/rust-clippy/master/index.html#assign_op_pattern)
* [redundant imports](https://rust-lang.github.io/rust-clippy/master/index.html#single_component_path_imports)
* [nonminimal_bool](https://rust-lang.github.io/rust-clippy/master/index.html#nonminimal_bool)
* [manual_memcpy](https://rust-lang.github.io/rust-clippy/master/index.html#manual_memcpy)
* [map_clone](https://rust-lang.github.io/rust-clippy/master/index.html#map_clone)
* [manual_strip](https://rust-lang.github.io/rust-clippy/master/index.html#manual_strip)